refactor: Rename primitive::basic to primitive::float
