+++ crates/polars-parquet/src/arrow/read/deserialize/primitive/float.rs
@@ -4,6 +4,10 @@ use arrow::datatypes::ArrowDataType;
 use arrow::types::NativeType;
 
 use super::super::utils;
+use super::{
+    deserialize_plain, AsDecoderFunction, ClosureDecoderFunction, DecoderFunction,
+    PlainDecoderFnCollector, PrimitiveDecoder, UnitDecoderFunction,
+};
 use crate::parquet::encoding::hybrid_rle::DictionaryTranslator;
 use crate::parquet::encoding::{byte_stream_split, hybrid_rle, Encoding};
 use crate::parquet::error::ParquetResult;
@@ -15,126 +19,6 @@ use crate::read::deserialize::utils::{
     TranslatedHybridRle,
 };
 
-/// A function that defines how to decode from the
-/// [`parquet::types::NativeType`][ParquetNativeType] to the [`arrow::types::NativeType`].
-///
-/// This should almost always be inlined.
-pub(crate) trait DecoderFunction<P, T>: Copy
-where
-    T: NativeType,
-    P: ParquetNativeType,
-{
-    fn decode(self, x: P) -> T;
-}
-
-#[derive(Default, Clone, Copy)]
-pub(crate) struct UnitDecoderFunction<T>(std::marker::PhantomData<T>);
-impl<T: NativeType + ParquetNativeType> DecoderFunction<T, T> for UnitDecoderFunction<T> {
-    #[inline(always)]
-    fn decode(self, x: T) -> T {
-        x
-    }
-}
-
-#[derive(Default, Clone, Copy)]
-pub(crate) struct AsDecoderFunction<P, T>(std::marker::PhantomData<(P, T)>);
-macro_rules! as_decoder_impl {
-    ($($p:ty => $t:ty,)+) => {
-        $(
-        impl DecoderFunction<$p, $t> for AsDecoderFunction<$p, $t> {
-            #[inline(always)]
-            fn decode(self, x : $p) -> $t {
-                x as $t
-            }
-        }
-        )+
-    };
-}
-
-as_decoder_impl![
-    i32 => i8,
-    i32 => i16,
-    i32 => u8,
-    i32 => u16,
-    i32 => u32,
-    i64 => i32,
-    i64 => u32,
-    i64 => u64,
-];
-
-#[derive(Default, Clone, Copy)]
-pub(crate) struct IntoDecoderFunction<P, T>(std::marker::PhantomData<(P, T)>);
-impl<P, T> DecoderFunction<P, T> for IntoDecoderFunction<P, T>
-where
-    P: ParquetNativeType + Into<T>,
-    T: NativeType,
-{
-    #[inline(always)]
-    fn decode(self, x: P) -> T {
-        x.into()
-    }
-}
-
-#[derive(Clone, Copy)]
-pub(crate) struct ClosureDecoderFunction<P, T, F>(F, std::marker::PhantomData<(P, T)>);
-impl<P, T, F> DecoderFunction<P, T> for ClosureDecoderFunction<P, T, F>
-where
-    P: ParquetNativeType,
-    T: NativeType,
-    F: Copy + Fn(P) -> T,
-{
-    #[inline(always)]
-    fn decode(self, x: P) -> T {
-        (self.0)(x)
-    }
-}
-
-pub(crate) struct PlainDecoderFnCollector<'a, 'b, P, T, D>
-where
-    T: NativeType,
-    P: ParquetNativeType,
-    D: DecoderFunction<P, T>,
-{
-    pub(crate) chunks: &'b mut ArrayChunks<'a, P>,
-    pub(crate) decoder: D,
-    pub(crate) _pd: std::marker::PhantomData<T>,
-}
-
-impl<'a, 'b, P, T, D: DecoderFunction<P, T>> BatchableCollector<(), Vec<T>>
-    for PlainDecoderFnCollector<'a, 'b, P, T, D>
-where
-    T: NativeType,
-    P: ParquetNativeType,
-    D: DecoderFunction<P, T>,
-{
-    fn reserve(target: &mut Vec<T>, n: usize) {
-        target.reserve(n);
-    }
-
-    fn push_n(&mut self, target: &mut Vec<T>, n: usize) -> ParquetResult<()> {
-        let n = usize::min(self.chunks.len(), n);
-        let (items, remainder) = self.chunks.bytes.split_at(n);
-        let decoder = self.decoder;
-        target.extend(
-            items
-                .iter()
-                .map(|chunk| decoder.decode(P::from_le_bytes(*chunk))),
-        );
-        self.chunks.bytes = remainder;
-        Ok(())
-    }
-
-    fn push_n_nulls(&mut self, target: &mut Vec<T>, n: usize) -> ParquetResult<()> {
-        target.resize(target.len() + n, T::default());
-        Ok(())
-    }
-
-    fn skip_in_place(&mut self, n: usize) -> ParquetResult<()> {
-        self.chunks.skip_in_place(n);
-        Ok(())
-    }
-}
-
 #[allow(clippy::large_enum_variant)]
 #[derive(Debug)]
 pub(crate) enum StateTranslation<'a, P: ParquetNativeType> {
@@ -143,7 +27,7 @@ pub(crate) enum StateTranslation<'a, P: ParquetNativeType> {
     ByteStreamSplit(byte_stream_split::Decoder<'a>),
 }
 
-impl<'a, P, T, D> utils::StateTranslation<'a, PrimitiveDecoder<P, T, D>> for StateTranslation<'a, P>
+impl<'a, P, T, D> utils::StateTranslation<'a, FloatDecoder<P, T, D>> for StateTranslation<'a, P>
 where
     T: NativeType,
     P: ParquetNativeType,
@@ -152,9 +36,9 @@ where
     type PlainDecoder = ArrayChunks<'a, P>;
 
     fn new(
-        _decoder: &PrimitiveDecoder<P, T, D>,
+        _decoder: &FloatDecoder<P, T, D>,
         page: &'a DataPage,
-        dict: Option<&'a <PrimitiveDecoder<P, T, D> as utils::Decoder>::Dict>,
+        dict: Option<&'a <FloatDecoder<P, T, D> as utils::Decoder>::Dict>,
         _page_validity: Option<&PageValidity<'a>>,
     ) -> ParquetResult<Self> {
         match (page.encoding(), dict) {
@@ -202,11 +86,11 @@ where
 
     fn extend_from_state(
         &mut self,
-        decoder: &mut PrimitiveDecoder<P, T, D>,
-        decoded: &mut <PrimitiveDecoder<P, T, D> as utils::Decoder>::DecodedState,
+        decoder: &mut FloatDecoder<P, T, D>,
+        decoded: &mut <FloatDecoder<P, T, D> as utils::Decoder>::DecodedState,
         is_optional: bool,
         page_validity: &mut Option<PageValidity<'a>>,
-        dict: Option<&'a <PrimitiveDecoder<P, T, D> as utils::Decoder>::Dict>,
+        dict: Option<&'a <FloatDecoder<P, T, D> as utils::Decoder>::Dict>,
         additional: usize,
     ) -> ParquetResult<()> {
         match self {
@@ -232,7 +116,7 @@ where
                     None => {
                         values.extend(
                             page_values
-                                .iter_converted(|v| decoder.decoder.decode(decode(v)))
+                                .iter_converted(|v| decoder.0.decoder.decode(decode(v)))
                                 .take(additional),
                         );
 
@@ -245,7 +129,7 @@ where
                         page_validity,
                         Some(additional),
                         values,
-                        &mut page_values.iter_converted(|v| decoder.decoder.decode(decode(v))),
+                        &mut page_values.iter_converted(|v| decoder.0.decoder.decode(decode(v))),
                     )?,
                 }
             },
@@ -256,17 +140,13 @@ where
 }
 
 #[derive(Debug)]
-pub(crate) struct PrimitiveDecoder<P, T, D>
+pub(crate) struct FloatDecoder<P, T, D>(PrimitiveDecoder<P, T, D>)
 where
     P: ParquetNativeType,
     T: NativeType,
-    D: DecoderFunction<P, T>,
-{
-    pub(crate) decoder: D,
-    _pd: std::marker::PhantomData<(P, T)>,
-}
+    D: DecoderFunction<P, T>;
 
-impl<P, T, D> PrimitiveDecoder<P, T, D>
+impl<P, T, D> FloatDecoder<P, T, D>
 where
     P: ParquetNativeType,
     T: NativeType,
@@ -274,14 +154,11 @@ where
 {
     #[inline]
     fn new(decoder: D) -> Self {
-        Self {
-            decoder,
-            _pd: std::marker::PhantomData,
-        }
+        Self(PrimitiveDecoder::new(decoder))
     }
 }
 
-impl<T> PrimitiveDecoder<T, T, UnitDecoderFunction<T>>
+impl<T> FloatDecoder<T, T, UnitDecoderFunction<T>>
 where
     T: NativeType + ParquetNativeType,
     UnitDecoderFunction<T>: Default + DecoderFunction<T, T>,
@@ -291,7 +168,7 @@ where
     }
 }
 
-impl<P, T> PrimitiveDecoder<P, T, AsDecoderFunction<P, T>>
+impl<P, T> FloatDecoder<P, T, AsDecoderFunction<P, T>>
 where
     P: ParquetNativeType,
     T: NativeType,
@@ -302,18 +179,7 @@ where
     }
 }
 
-impl<P, T> PrimitiveDecoder<P, T, IntoDecoderFunction<P, T>>
-where
-    P: ParquetNativeType,
-    T: NativeType,
-    IntoDecoderFunction<P, T>: Default + DecoderFunction<P, T>,
-{
-    pub(crate) fn cast_into() -> Self {
-        Self::new(IntoDecoderFunction::<P, T>::default())
-    }
-}
-
-impl<P, T, F> PrimitiveDecoder<P, T, ClosureDecoderFunction<P, T, F>>
+impl<P, T, F> FloatDecoder<P, T, ClosureDecoderFunction<P, T, F>>
 where
     P: ParquetNativeType,
     T: NativeType,
@@ -330,7 +196,7 @@ impl<T> utils::ExactSize for (Vec<T>, MutableBitmap) {
     }
 }
 
-impl<P, T, D> utils::Decoder for PrimitiveDecoder<P, T, D>
+impl<P, T, D> utils::Decoder for FloatDecoder<P, T, D>
 where
     T: NativeType,
     P: ParquetNativeType,
@@ -349,7 +215,7 @@ where
     }
 
     fn deserialize_dict(&self, page: DictPage) -> ParquetResult<Self::Dict> {
-        Ok(deserialize_plain::<P, T, D>(&page.buffer, self.decoder))
+        Ok(deserialize_plain::<P, T, D>(&page.buffer, self.0.decoder))
     }
 
     fn decode_plain_encoded<'a>(
@@ -364,7 +230,7 @@ where
             None => {
                 PlainDecoderFnCollector {
                     chunks: page_values,
-                    decoder: self.decoder,
+                    decoder: self.0.decoder,
                     _pd: std::marker::PhantomData,
                 }
                 .push_n(values, limit)?;
@@ -376,7 +242,7 @@ where
             Some(page_validity) => {
                 let collector = PlainDecoderFnCollector {
                     chunks: page_values,
-                    decoder: self.decoder,
+                    decoder: self.0.decoder,
                     _pd: std::marker::PhantomData,
                 };
 
@@ -439,7 +305,7 @@ where
     }
 }
 
-impl<P, T, D> utils::DictDecodable for PrimitiveDecoder<P, T, D>
+impl<P, T, D> utils::DictDecodable for FloatDecoder<P, T, D>
 where
     T: NativeType,
     P: ParquetNativeType,
@@ -462,7 +328,7 @@ where
     }
 }
 
-impl<P, T, D> utils::NestedDecoder for PrimitiveDecoder<P, T, D>
+impl<P, T, D> utils::NestedDecoder for FloatDecoder<P, T, D>
 where
     T: NativeType,
     P: ParquetNativeType,
@@ -484,17 +350,4 @@ where
     ) {
         values.resize(values.len() + n, T::default());
     }
-}
-
-pub(super) fn deserialize_plain<P, T, D>(values: &[u8], decoder: D) -> Vec<T>
-where
-    T: NativeType,
-    P: ParquetNativeType,
-    D: DecoderFunction<P, T>,
-{
-    values
-        .chunks_exact(std::mem::size_of::<P>())
-        .map(decode)
-        .map(|v| decoder.decode(v))
-        .collect::<Vec<_>>()
 }
